# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RoadOps API
  version: 1.0.0
  description: >-
    Yo'l ekspluatatsiyasini boshqarish tizimining brauzer va integratsiya API si.
    RoadVision yozuvi faqat ko'rik navbatiga tushadi; inson tasdig'isiz nuqson yoki
    reja yaratilmaydi. API ustuvorlik balli, holat indeksi, 0-100 bahosi yoki AI
    ishonch foizini qabul qilmaydi va qaytarmaydi. Operatsion ko'lam autentifikatsiya
    qilingan foydalanuvchining yo'l bo'limi va unga tegishli YTP yo'llari bilan cheklanadi.
  contact:
    name: RoadOps platform team
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Same-origin web gateway
...
paths:
...
  /roadvision/findings:
    get:
      tags: [Findings]
      summary: List RoadVision observations
      operationId: listRoadVisionFindings
      parameters:
        - name: state
          in: query
          required: false
          schema:
            allOf:
              - $ref: "#/components/schemas/FindingState"
            default: PENDING_REVIEW
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Vendor observations in the human review lifecycle.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PagedRoadVisionFindings"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/RoadScopeUnavailable" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: roadops_session
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  responses:
    Unauthenticated:
      description: Session is absent, expired or revoked.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Authenticated user lacks the permission or road-unit scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    RoadScopeUnavailable:
      description: >-
        Operational processing is fail-closed because the requested road is missing, ambiguous,
        inactive or outside the authenticated actor's division scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ValidationError:
      description: Request cannot be processed as supplied.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          type: object
          additionalProperties: false
          properties:
            requestId: { type: string, format: uuid }
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string, pattern: "^[A-Z][A-Z0-9_]*$" }
            message: { type: string, minLength: 1 }
            details:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
            requestId: { type: string, format: uuid }
      additionalProperties: false
...
    DivisionRef:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1 }
    RoadRef:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: { type: string, minLength: 1, maxLength: 50 }
        name: { type: string, minLength: 1 }
    Quantity:
      type: object
      additionalProperties: false
      required: [value, unit]
      properties:
        value: { type: string, pattern: "^[0-9]+(?:\\.[0-9]+)?$" }
        unit: { type: string, minLength: 1, maxLength: 30 }
...
    FindingState:
      type: string
      enum: [PENDING_REVIEW, VERIFIED, REJECTED, DUPLICATE]
    RoadVisionFinding:
      type: object
      additionalProperties: false
      required:
        - id
        - vendorReference
        - attributeName
        - road
        - division
        - chainageStartM
        - observedAt
        - receivedAt
        - state
        - evidence
      properties:
        id: { type: string, format: uuid }
        vendorReference: { type: string, minLength: 1 }
        attributeName: { type: string, minLength: 1 }
        road: { $ref: "#/components/schemas/RoadRef" }
        division: { $ref: "#/components/schemas/DivisionRef" }
        chainageStartM: { type: number, minimum: 0 }
        chainageEndM: { type: number, minimum: 0 }
        laneLabel: { type: [string, "null"], maxLength: 100 }
        observedAt: { type: string, format: date-time }
        receivedAt: { type: string, format: date-time }
        state: { $ref: "#/components/schemas/FindingState" }
        measuredQuantity: { $ref: "#/components/schemas/Quantity" }
        evidence:
          type: array
          maxItems: 100
          description: Every complete vendor media entry, with no private object URI.
          items: { $ref: "#/components/schemas/RoadVisionEvidenceMedia" }
        reviewerNote: { type: string, maxLength: 5000 }
    RoadVisionEvidenceMedia:
      type: object
      additionalProperties: false
      required: [index, contentType, capturedAt, sha256, url]
      properties:
        index: { type: integer, minimum: 0, maximum: 99 }
        mediaId: { type: [string, "null"], maxLength: 200 }
        contentType: { type: string, enum: [image/jpeg, image/png, video/mp4] }
        capturedAt: { type: string, format: date-time }
        sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
        url:
          type: string
          pattern: "^/api/v1/roadvision/findings/[0-9a-fA-F-]{36}/evidence/(?:0|[1-9][0-9]?)$"
    PagedRoadVisionFindings:
      $ref: "#/components/schemas/PagedCollectionRoadVisionFinding"
    PagedCollectionRoadVisionFinding:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/RoadVisionFinding" }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
...
security:
  - sessionCookie: []